sorts renames modals
